Move load_station_data where it belongs
This now completely separates data loading/saving from frontend
